Scroll Rect for Story Text not working
Fixed logging of changing states, but switching to Scroll Rect did not work.  Doing so somehow put the Interaction Text (formerly Story Area Text) field behind everything so it couldn't be seen, but could still be written to.  Split the Story Text Background so the it only is for Story Text and created new Interaction Text background, but it still covers up the Interaction Text field so it can't be seen.