First working web version
Seems to have a problem with the image used for the scrollable mask in the top text area.  When run as a local web page, it looks fine in full screen mode, but not on the https://sharemygame.com/@lvskiprof/warlock website, even in full screen mode.  It is wrong in anything less than full screen mode locally.